Add Alma payment method with new payment API fields
Moved the address fields to the parent object
Now we can declare new fields dinamically as long as they follow create+$fieldName
