+++ src/PaymentMethods/Alma.php
@@ -22,6 +22,10 @@ class Alma extends AbstractPaymentMethod implements PaymentMethodI
             'filtersOnBuild' => false,
             'confirmationDelayed' => false,
             'SEPA' => false,
+            'paymentAPIfields' => [
+                'billingAddress',
+                'shippingAddress',
+            ],
         ];
     }
