+++ src/Shared/SharedDataDictionary.php
@@ -34,6 +34,7 @@ class SharedDataDictionary
         'Mollie_WC_Gateway_Blik',
         'Mollie_WC_Gateway_Twint',
         'Mollie_WC_Gateway_Bancomatpay',
+        'Mollie_WC_Gateway_Alma',
     ];
 
     public const MOLLIE_OPTIONS_NAMES = [
